Move api to :common
